requirements.txt
python-dotenv
fastapi
uvicorn
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
groq
langserve
